app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
qdrant-client>=1.7.0
python-dotenv>=0.19.0
loguru>=0.6.0
httpx>=0.23.0
uuid>=1.30
jinja2>=3.0.0